update bezout_2.py & tex/txt/

+++ python/bezout_2.py
@@ -122,14 +122,16 @@ def P2txt(n, deg, P, directory):
 
 def bz2txt(n, directory, BB):
     for f in os.listdir(directory+'/BB'):
-        os.remove(directory+'/BB/'+f)
+		if f[-4:] == '.txt':
+			os.remove(directory+'/BB/'+f)
     for k in range(n+1):
         m = BB[k]
         mat2txt(m, directory+'/BB/bb'+str(k))
 
 def gb2txt(directory, GB):
 	for f in os.listdir(directory+'/GB'):
-		os.remove(directory+'/GB/'+f)
+		if f[-4:] == '.txt':
+			os.remove(directory+'/GB/'+f)
 	for k in range(len(GB)):
 		p = GB[k]
 		mat2txt(p, directory+'/GB/gb'+str(k))
